Refactoring consideration, still determining if it's the right path
